[refac]: renamed files according to PEP8
